move binary matrix func and lint
