+++ pennylane/pauli/utils.py
@@ -1247,6 +1247,7 @@ def diagonalize_qwc_groupings(qwc_groupings):
 
     return post_rotations, diag_groupings
 
+
 # from observable_hf.py -------------------------
 def simplify(h, cutoff=1.0e-12):
     r"""Add together identical terms in the Hamiltonian.
@@ -1377,3 +1378,47 @@ pauli_coeff = {
     "XY": 1.0j,
     "YX": -1.0j,
 }
+
+
+# from tapering.py ------------------------------
+def _binary_matrix(terms, num_qubits, wire_map=None):
+    r"""Get a binary matrix representation of the Hamiltonian where each row corresponds to a
+    Pauli term, which is represented by a concatenation of Z and X vectors.
+
+    Args:
+        terms (Iterable[Observable]): operators defining the Hamiltonian
+        num_qubits (int): number of wires required to define the Hamiltonian
+        wire_map (dict): dictionary containing all wire labels used in the Pauli words as keys, and
+            unique integer labels as their values
+
+    Returns:
+        array[int]: binary matrix representation of the Hamiltonian of shape
+        :math:`len(terms) \times 2*num_qubits`
+
+    **Example**
+
+    >>> wire_map = {'a':0, 'b':1, 'c':2, 'd':3}
+    >>> terms = [qml.PauliZ(wires=['a']) @ qml.PauliX(wires=['b']),
+    ...          qml.PauliZ(wires=['a']) @ qml.PauliY(wires=['c']),
+    ...          qml.PauliX(wires=['a']) @ qml.PauliY(wires=['d'])]
+    >>> _binary_matrix(terms, 4, wire_map=wire_map)
+    array([[1, 0, 0, 0, 0, 1, 0, 0],
+           [1, 0, 1, 0, 0, 0, 1, 0],
+           [0, 0, 0, 1, 1, 0, 0, 1]])
+    """
+    if wire_map is None:
+        all_wires = qml.wires.Wires.all_wires([term.wires for term in terms], sort=True)
+        wire_map = {i: c for c, i in enumerate(all_wires)}
+
+    binary_matrix = np.zeros((len(terms), 2 * num_qubits), dtype=int)
+    for idx, term in enumerate(terms):
+        ops, wires = term.name, term.wires
+        if len(wires) == 1:
+            ops = [ops]
+        for op, wire in zip(ops, wires):
+            if op in ["PauliX", "PauliY"]:
+                binary_matrix[idx][wire_map[wire] + num_qubits] = 1
+            if op in ["PauliZ", "PauliY"]:
+                binary_matrix[idx][wire_map[wire]] = 1
+
+    return binary_matrix
